Update: devide all trips and trip card components add redux
